requirements.txt
azure-functions
python-dotenv
azure.cosmos
pydantic
rapidfuzz
langchain
langchain_openai
langgraph
azure-ai-documentintelligence
azure-cosmos            
azure-functions         
azure-functions-durable 
azure-identity          
azure-storage-blob      
langchain               
langchain-community     
langchain-core          
langchain-openai        
langchain-text-splitters
langgraph               
langgraph-checkpoint    
pydantic                
pydantic-settings       
pydantic_core           
Pygments                
PyJWT                   
PyMuPDF                 
pyparsing               
pypdf                   
PyPDF2                  
pytesseract             
python-dateutil         
python-dotenv           
requests
azure-ai-projects 
azure-ai-evaluation
azure-ai-agents
azure-identity-broker
pandas
# azure.ai.contentsafety
parameterized
openpyxl
jinja2
langchain-classic
python-tools